+++ lib/cocoapods/generator/app_target_helper.rb
@@ -99,6 +99,7 @@ module Pod
       # @return [void]
       #
       def self.add_swift_version(target, swift_version)
+        raise 'Cannot set empty Swift version to target.' if swift_version.blank?
         target.build_configurations.each do |configuration|
           configuration.build_settings['SWIFT_VERSION'] = swift_version
         end